move file structure less complex
