--- scripts/index.js
@@ -1,8 +1,0 @@
-#!/usr/bin/env node
-/* eslint-disable import/no-commonjs */
-
-// instantsearch.js optional-name-of-transform optional/path/to/src [...options]
-
-const { run } = require('./transforms');
-
-run();
