+++ scripts/transforms/script.js
@@ -1,0 +1,244 @@
+/*
+ * This code is derived from https://github.com/reactjs/react-codemod, which has the following license:
+ *
+ * Copyright (c) 2015-present, Facebook, Inc.
+ *
+ * Licensed under the MIT license.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+/* eslint-disable no-process-exit,import/no-commonjs,no-console */
+const path = require('path');
+
+const globby = require('globby');
+const inquirer = require('inquirer');
+const meow = require('meow');
+const execa = require('execa');
+const isGitClean = require('is-git-clean');
+
+const jscodeshiftExecutable = require.resolve('.bin/jscodeshift');
+const transformsDirectory = path.join(__dirname, './');
+
+const YELLOW = '\x1b[33m%s\x1b[0m';
+
+function checkGitStatus(force) {
+  let clean = false;
+  let errorMessage = 'Unable to determine if git directory is clean';
+  try {
+    clean = isGitClean.sync(process.cwd());
+    errorMessage = 'Git directory is not clean';
+  } catch (err) {
+    if (err && err.stderr && err.stderr.indexOf('Not a git repository') >= 0) {
+      clean = true;
+    }
+  }
+
+  if (!clean) {
+    if (force) {
+      console.log(`WARNING: ${errorMessage}. Forcibly continuing.`);
+    } else {
+      console.log('Thank you for using InstantSearch codemods!');
+      console.log(
+        YELLOW,
+        '\nBut before we continue, please stash or commit your git changes.'
+      );
+      console.log(
+        '\nYou may use the --force flag to override this safety check.'
+      );
+      process.exit(1);
+    }
+  }
+}
+
+function runTransform({ files, flags, parser, transformer }) {
+  const transformerPath = path.join(transformsDirectory, `${transformer}.js`);
+
+  let args = [];
+
+  const { dry, print, explicitRequire } = flags;
+
+  if (dry) {
+    args.push('--dry');
+  }
+  if (print) {
+    args.push('--print');
+  }
+
+  if (explicitRequire === 'false') {
+    args.push('--explicit-require=false');
+  }
+
+  args.push('--verbose=2');
+
+  args.push('--ignore-pattern=**/node_modules/**');
+
+  args.push('--parser', parser);
+
+  if (parser === 'tsx') {
+    args.push('--extensions=tsx,ts,jsx,js');
+  } else {
+    args.push('--extensions=jsx,js');
+  }
+
+  args = args.concat(['--transform', transformerPath]);
+
+  if (flags.jscodeshift) {
+    args = args.concat(flags.jscodeshift);
+  }
+
+  args = args.concat(files);
+
+  console.log(`Executing command: jscodeshift ${args.join(' ')}`);
+
+  const result = execa.sync(jscodeshiftExecutable, args, {
+    stdio: 'inherit',
+    stripEof: false,
+  });
+
+  if (result.error) {
+    throw result.error;
+  }
+}
+
+const TRANSFORMER_INQUIRER_CHOICES = [
+  {
+    name:
+      'addWidget-to-addWidgets: Replace usage of addWidget and removeWidget to their plural form',
+    value: 'addWidget-to-addWidgets',
+  },
+];
+
+const PARSER_INQUIRER_CHOICES = [
+  {
+    name: 'JavaScript',
+    value: 'babel',
+  },
+  {
+    name: 'JavaScript with Flow',
+    value: 'flow',
+  },
+  {
+    name: 'TypeScript',
+    value: 'tsx',
+  },
+];
+
+function expandFilePathsIfNeeded(filesBeforeExpansion) {
+  const shouldExpandFiles = filesBeforeExpansion.some(file =>
+    file.includes('*')
+  );
+  return shouldExpandFiles
+    ? globby.sync(filesBeforeExpansion)
+    : filesBeforeExpansion;
+}
+
+function run() {
+  const cli = meow(
+    {
+      description: 'Codemods for updating InstantSearch usage',
+      help: `
+    Usage
+      $ npx instantsearch.js <transform> <path> <...options>
+        transform    One of the choices from https://github.com/instantsearch.js/ 
+        path         Files or directory to transform. Can be a glob like src/**.test.js
+    Options
+      --force        Bypass Git safety checks and forcibly run codemods
+      --dry          Dry run (no changes are made to files)
+      --print        Print transformed files to your terminal
+      --jscodeshift  (Advanced) Pass options directly to jscodeshift
+    `,
+    },
+    {
+      boolean: ['force', 'dry', 'print', 'explicit-require', 'help'],
+      string: ['_'],
+      alias: {
+        h: 'help',
+      },
+    }
+  );
+
+  if (!cli.flags.dry) {
+    checkGitStatus(cli.flags.force);
+  }
+
+  if (
+    cli.input[0] &&
+    !TRANSFORMER_INQUIRER_CHOICES.find(x => x.value === cli.input[0])
+  ) {
+    console.error('Invalid transform choice, pick one of:');
+    console.error(
+      TRANSFORMER_INQUIRER_CHOICES.map(x => `- ${x.value}`).join('\n')
+    );
+    process.exit(1);
+  }
+
+  inquirer
+    .prompt([
+      {
+        type: 'input',
+        name: 'files',
+        message: 'On which files or directory should the codemods be applied?',
+        when: !cli.input[1],
+        default: '.',
+        filter: files => files.trim(),
+      },
+      {
+        type: 'list',
+        name: 'parser',
+        message: 'Which dialect of JavaScript do you use?',
+        default: 'babel',
+        when: !cli.flags.parser,
+        pageSize: PARSER_INQUIRER_CHOICES.length,
+        choices: PARSER_INQUIRER_CHOICES,
+      },
+      {
+        type: 'list',
+        name: 'transformer',
+        message: 'Which transform would you like to apply?',
+        when: !cli.input[0],
+        pageSize: TRANSFORMER_INQUIRER_CHOICES.length,
+        choices: TRANSFORMER_INQUIRER_CHOICES,
+      },
+    ])
+    .then(answers => {
+      const { files, transformer, parser } = answers;
+
+      const filesBeforeExpansion = cli.input[1] || files;
+      const filesExpanded = expandFilePathsIfNeeded([filesBeforeExpansion]);
+
+      const selectedTransformer = cli.input[0] || transformer;
+      const selectedParser = cli.flags.parser || parser;
+
+      if (!filesExpanded.length) {
+        console.log(
+          `No files found matching ${filesBeforeExpansion.join(' ')}`
+        );
+        return null;
+      }
+
+      return runTransform({
+        files: filesExpanded,
+        flags: cli.flags,
+        parser: selectedParser,
+        transformer: selectedTransformer,
+      });
+    });
+}
+
+module.exports = { run };
